+++ source/Core/Module/Module.php
@@ -428,7 +428,6 @@ class Module extends \OxidEsales\Eshop\Core\Base
         $data[MetaDataProvider::METADATA_CONTROLLERS] = $this->convertControllersToArray($moduleConfiguration);
         $data[MetaDataProvider::METADATA_SMARTY_PLUGIN_DIRECTORIES] =
             $this->convertSmartyPluginDirectoriesToArray($moduleConfiguration);
-        $data[MetaDataProvider::METADATA_BLOCKS] = $this->convertTemplateBlocksToArray($moduleConfiguration);
         $data[MetaDataProvider::METADATA_EVENTS] = $this->convertEventsToArray($moduleConfiguration);
         $data[MetaDataProvider::METADATA_SETTINGS] = $this->convertSettingsToArray($moduleConfiguration);
 
@@ -498,30 +497,6 @@ class Module extends \OxidEsales\Eshop\Core\Base
 
         foreach ($moduleConfiguration->getControllers() as $controller) {
             $data[$controller->getId()] = $controller->getControllerClassNameSpace();
-        }
-
-        return $data;
-    }
-    /**
-     * @param ModuleConfiguration $moduleConfiguration
-     * @return array
-     */
-    private function convertTemplateBlocksToArray(ModuleConfiguration $moduleConfiguration): array
-    {
-        $data = [];
-
-        foreach ($moduleConfiguration->getTemplateBlocks() as $key => $templateBlock) {
-            $data[$key] = [
-                'template' => $templateBlock->getShopTemplatePath(),
-                'block' => $templateBlock->getBlockName(),
-                'file' => $templateBlock->getModuleTemplatePath(),
-            ];
-            if ($templateBlock->getTheme() !== '') {
-                $data[$key]['theme'] = $templateBlock->getTheme();
-            }
-            if ($templateBlock->getPosition() !== 0) {
-                $data[$key]['position'] = $templateBlock->getPosition();
-            }
         }
 
         return $data;