+++ source/Internal/Framework/Module/Configuration/DataObject/ModuleConfiguration.php
@@ -14,7 +14,6 @@ use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject\ModuleConfiguration\Event;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject\ModuleConfiguration\SmartyPluginDirectory;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject\ModuleConfiguration\Template;
-use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject\ModuleConfiguration\TemplateBlock;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\Exception\ModuleSettingNotFountException;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Setting\Setting;
 
@@ -88,11 +87,6 @@ class ModuleConfiguration
      * @var SmartyPluginDirectory[]
      */
     private $smartyPluginDirectories = [];
-
-    /**
-     * @var TemplateBlock[]
-     */
-    private $templateBlocks = [];
 
     /**
      * @var Event[]
@@ -345,29 +339,6 @@ class ModuleConfiguration
     public function hasClassExtensions(): bool
     {
         return !empty($this->classExtensions);
-    }
-
-    /**
-     * @return TemplateBlock[]
-     */
-    public function getTemplateBlocks(): array
-    {
-        return $this->templateBlocks;
-    }
-
-    public function addTemplateBlock(TemplateBlock $templateBlock)
-    {
-        $this->templateBlocks[] = $templateBlock;
-
-        return $this;
-    }
-
-    /**
-     * @return bool
-     */
-    public function hasTemplateBlocks(): bool
-    {
-        return !empty($this->templateBlocks);
     }
 
     /**