+++ source/Internal/Framework/Smarty/Module/TemplateExtension/TemplateBlockExtensionServiceInterface.php
@@ -7,7 +7,7 @@
 
 declare(strict_types=1);
 
-namespace OxidEsales\EshopCommunity\Internal\Framework\Module\TemplateExtension;
+namespace OxidEsales\EshopCommunity\Internal\Framework\Smarty\Module\TemplateExtension;
 
 interface TemplateBlockExtensionServiceInterface
 {