+++ source/Internal/Framework/Smarty/Plugin/prefilter.oxblock.php
@@ -6,7 +6,7 @@
  */
 
 use OxidEsales\EshopCommunity\Internal\Container\ContainerFactory;
-use OxidEsales\EshopCommunity\Internal\Framework\Module\TemplateExtension\TemplateBlockExtensionServiceInterface;
+use OxidEsales\EshopCommunity\Internal\Framework\Smarty\Module\TemplateExtension\TemplateBlockExtensionServiceInterface;
 
 
 /**