+++ tests/Integration/Internal/Framework/Module/Configuration/Dao/ProjectConfigurationDaoTest.php
@@ -17,7 +17,6 @@ use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject\ModuleConfiguration;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject\ModuleConfiguration\Template;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject\ModuleConfiguration\SmartyPluginDirectory;
-use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject\ModuleConfiguration\TemplateBlock;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject\ProjectConfiguration;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Setting\Setting;
 use OxidEsales\EshopCommunity\Tests\ContainerTrait;
@@ -123,13 +122,6 @@ class ProjectConfigurationDaoTest extends TestCase
 
     private function getTestProjectConfiguration(): ProjectConfiguration
     {
-        $templateBlock = new TemplateBlock(
-            'extendedTemplatePath',
-            'testBlock',
-            'filePath'
-        );
-        $templateBlock->setTheme('flow_theme');
-        $templateBlock->setPosition(3);
         $moduleConfiguration = new ModuleConfiguration();
         $moduleConfiguration
             ->setId('testModuleConfiguration')
@@ -172,7 +164,6 @@ class ProjectConfigurationDaoTest extends TestCase
                     'secondSmartyDirectory'
                 )
             )
-            ->addTemplateBlock($templateBlock)
             ->addClassExtension(
                 new ClassExtension(
                     'originalClassNamespace',